增加四条协议

+++ module/User/src/User/Service/UserService.php
@@ -373,6 +373,7 @@ class UserService implements ServiceManagerAwareInterface, LoggerAwareInterface
 
             if ($curFullPath != '')
             {
+                if (file_exists($curFullPath))
                 unlink($curFullPath);
             }
